Separate app_state from AppState class.
Moved from_yaml method from config.py to global_app_state.py.
A default filename config.sample.yml is supplied to app_state for convenience.
