+++ backend/app/core/app_state.py
@@ -1,0 +1,3 @@
+from app.core.global_app_state import AppState
+
+app_state = AppState("config.sample.yml")
