+++ backend/app/core/config.py
@@ -49,7 +49,6 @@ class Settings(BaseSettings):
     def validate_matrix_bot(cls, v, values):
         return MatrixBotSettings.parse_obj(v)
 
-    @classmethod
     def from_yaml(cls, path_to_file):
         absolute_path_to_file = Path(path_to_file).absolute()
         try:
@@ -58,6 +57,3 @@ class Settings(BaseSettings):
                 return Settings.parse_obj(yaml_settings)
         except (IOError, ImportError) as err:
             print(f"Couldn't load config from file. Error: {err}")
-
-
-settings = Settings.from_yaml("config.yml")
